requirements.txt
torch
torchvision
pillow
pandas
scikit-learn
langdetect
tqdm
